Initial Commit. Contains entire project so far over the past month or so.
External libraries used:
-SDL 2
-SDL 2 Mixer (Audio)
-OpenGL
-GL Maths Library (GLM)

Currently in-place:
-Shader Loader
Simply loads shaders and adds them to a registry on creation. Features and updateShader funtion to allow shader manipulation at runtime. (Necessary for drawing an object at multiple different places during the same render cycle concurrently efficiently.)

-Model Loader
.exm files. Made them up myself because I'm awkward and wanted to try something different. Turns out that the file layout is rather similar to .obj in essence.

-3D Audio Simulation
Simulates audio around a player entity. Had to write custom "audio program" for SDL_mixer to allow panning as the pan audio functions with SDL_mixer do not work correctly. I still wish to apply elements from the Haas effect with the audio in the future as well as adding a slight delay to the audio if the audio source is played behind the player in order to build a better sense of direction.

-Entities
Objects with a postition. Child classes of this are:
--EntityUpdateable, which updates every second.
--EntityPlayer (From EntityUpdateable), which has a camera pitch and yaw.
--EntityAudioSource (From EntityUpdateable), which is used for 3D audio.