Attempting to fix camera movement.
Just playing around with it now, attempting to get strafing working correctly. I will continue to work on smooth camera movements and a better input system in updates following this.

Also made the external libraries more accessable and "plug 'n' play". As I can't block the .vcxproj file from being uploaded without causing problems with the project (that is the project), this is the alternative. The reason I state this is because the external directories are included with the .vcxproj. Now, it should be easier to load the project into new computers without too much adjusting for each push.

Just make sure to have the following:

--C:
----\Programming Libraries
------\glew
------\glm
------\SDL2
------\Simple OpenGL Image Library

Please also make sure to have added the SDL2_mixer library to SDL2.

And, of course, this only applies to those wishing to edit the project in Visual Studio and are running on Windows 10. (Other versions of Windows may work.)