-Added an input interface to make cleaning up the project easier.
-Cleaned up EntityUpdateable.
--Removed all speed calculation code as it's not needed yet.
--Changed how position data works.
---Before, a variable that stored the position to be updated was reset every time the position needed to be changed. However, when taking the player's movement, for example, they might need to walk with a forward vector and a left movement vector.
---Now, the new position is updated and changed. On update, the new position is applied.

////////

My aim is to try and rid Main.cpp of any temporary code or any code that directly interfaces with any specific element via abstraction. I don't want to go overboard with this, as I intend to begin to implement the beginnings of a 'game' at some point in the near future, but I want to try and do this to some degree. I also want to clean up the codebase and make it easier to understand in-case I decide to bring more people into the project at some point. It will also make the 'game' element of developement easier to work on, as more effort can be placed on other areas of production.

After all, this is just a very basic engine at the moment. This is early days and much is still left to be completed.