-Moved render functions out of main.cpp. This proved to be highly problematic. Had to resort to making the local player an external variable and all of the render functions as a portable class. There was a strange instance where shader programs were getting deleted. I soon discovered that this was due to the shader program variables held in Render.h to be deleted whenever the constructor was deleted. I did a slightly hacky work-around and made the shader program variables pointers. This has taken the better part of 5 hours to fix.
-Preparation for adding UI elements.
 (Specifically, a console interface.)
--Created console UI model and shader.
--Added in logic to lock/unlock mouse.

I'm preparing this code to add the following:
-Texture Loading via SOIL
-Bitmap Text Renderer (Fonts are boring.)
-UI Elements & Console

Once I have enabled this, I should be able to continue working on the game in fullscreen. Not overly important at this stage, but it would just be nice (and convenient!) to have.

After I have this basis in place, I want to begin working on gameplay elements. However, I think it would be a wise idea to introduce networking early on, just to prevent hassle in the future.
-Networking (Likely think.async / BoostASIO)
-Player instancing.
-BASIC!!!! Physics
-Inventory -- Items, etc.

Immediately after this, I want to sort out the graphical elements to a 'playable' standard.
-Skybox.
-Lighting.
-Nice(er) UI.

After this, I want to work on a plugin for Blender to allow me to save out to my .exm format. I don't think this should be too difficult. .exm and .obj are very similar so perhaps I should start looking there.

Once I can easily export out of a pre-established 3D modelling program, I want to begin working on designing levels and an editor. Considering I'm wanting to build an MMO or an Online RPG of some description, I simply cannot hard-code the environment in. I'll need to work on some way of editing the world. It just depends on whether I decide to do that with this program via a debugging mode of some description (after all, I'll have an editor) or whether I should create a completely seperate program. I'll have to see.

I could also go down the route of procedural generation, but I want to make sure there's a definite storyline so this could prove to be a challenge in a completely different regard. I just want every player's game to be unique one way or another. For instance, I want each player to *be able* to buy their own house and have the house be physical instead of it being instanced.

I don't know. I'll just make sure I'm tackling one thing at a time.